[refactor] PermissionDialog core:ui 모듈로 이동
GoToAppSetting -> NavigateToAppSetting 네이밍 변경(화면 이동은 전부 NavigateToDestination 형태로)
