+++ core/common/src/main/kotlin/com/unifest/android/core/common/PermissionDialogButtonType.kt
@@ -1,0 +1,7 @@
+package com.unifest.android.core.common
+
+enum class PermissionDialogButtonType {
+    DISMISS,
+    CONFIRM,
+    NAVIGATE_TO_APP_SETTING,
+}
