+++ feature/map/src/main/kotlin/com/unifest/android/feature/map/viewmodel/MapUiEvent.kt
@@ -4,7 +4,7 @@ import com.unifest.android.core.common.UiText
 
 sealed interface MapUiEvent {
     data object RequestLocationPermission : MapUiEvent
-    data object GoToAppSettings : MapUiEvent
+    data object NavigateToAppSetting : MapUiEvent
     data class NavigateToBoothDetail(val boothId: Long) : MapUiEvent
     data class ShowSnackBar(val message: UiText) : MapUiEvent
 }
